Make minor modifications
