+++ charon/src/graphs.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 use std::iter::FromIterator;
 use std::vec::Vec;
 
-/// A structure containing information about SCCs (Strongly Connex Components)
+/// A structure containing information about SCCs (Strongly Connected Components)
 pub struct SCCs<Id> {
     /// The SCCs themselves
     pub sccs: Vec<Vec<Id>>,
@@ -68,7 +68,7 @@ fn insert_scc_with_deps<Id: Copy + std::hash::Hash + Eq>(
 
 /// Provided we computed the SCCs (Strongly Connected Components) of a set of
 /// identifier, and those identifiers are ordered, compute the set of SCCs where
-/// the order of the SCCs and the order of the identifiers inside the SCCs try
+/// the order of the SCCs and the order of the identifiers inside the SCCs attempt
 /// to respect as much as possible the original order between the identifiers.
 /// The `ids` vector gives the ordered set of identifiers.
 ///
