+++ tests/src/lib.rs
@@ -2,6 +2,7 @@ mod constants;
 mod external;
 mod hashmap;
 mod loops;
+mod loops_cfg;
 mod matches;
 mod nested_borrows;
 mod no_nested_borrows;
